# Conflicts:
#	src/main/java/com/example/parkinglot/entity/Car.java
#	src/main/java/com/example/parkinglot/repo/CarRepository.java
#	src/main/java/com/example/parkinglot/service/CarService.java
#	src/main/webapp/app/admin-sidebar/admin-sidebar.component.html
#	src/main/webapp/app/app.routes.ts
#	src/main/webapp/app/car/car.model.ts
#	src/main/webapp/app/car/delete/car-delete-dialog.component.html
#	src/main/webapp/app/car/delete/car-delete-dialog.component.spec.ts
#	src/main/webapp/app/car/delete/car-delete-dialog.component.ts
#	src/main/webapp/app/car/detail/car-detail.component.html
#	src/main/webapp/app/car/detail/car-detail.component.spec.ts
#	src/main/webapp/app/car/detail/car-detail.component.ts
#	src/main/webapp/app/car/list/car.component.html
#	src/main/webapp/app/car/list/car.component.spec.ts
#	src/main/webapp/app/car/list/car.component.ts
#	src/main/webapp/app/car/service/car.service.spec.ts
#	src/main/webapp/app/car/service/car.service.ts
#	src/main/webapp/app/car/update/car-update.component.html
#	src/main/webapp/app/car/update/car-update.component.spec.ts
#	src/main/webapp/app/car/update/car-update.component.ts

+++ src/main/java/com/example/parkinglot/entity/Price.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.Data;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 
 @Entity
 @Data
@@ -12,7 +11,7 @@ import java.math.BigInteger;
 public class Price {
 
     @Id
-    private int duration;
+    private Long duration;
 
     private BigDecimal price;
 }
